# backend/static/api.yaml
openapi: 3.0.0
info:
  title: ParkMan API
  description: A REST API for booking and managing parking lots
  version: 1.0.0
servers:
  - url: http://localhost:5000
  # description: Local Development Server
paths:
...
  /api/parking-lots/:
    get:
      summary: Get parking lots
      description: Allows users to fetch all parking lots from the database.
      tags:
        - Parking Lots
      responses:
        "200":
          description: Successfully fetched parking lots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  lots:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        prime_location_name:
                          type: string
                        address:
                          type: string
                        pincode:
                          type: string
                        number_of_spots:
                          type: string
              example:
                success: true
                lots:
                  - id: 1
                    prime_location_name: "Central Mall"
                    address: "123 Main Street"
                    pincode: "560001"
                    price: 50.00
                    number_of_spots: 120
                  - id: 2
                    prime_location_name: "Tech Park"
                    address: "456 Silicon Avenue"
                    pincode: "560102"
                    price: 40.00
                    number_of_spots: 90

        "404":
          description: Failed to fetch parking lots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: false
                message: No parking lots found
        "500":
          description: Failed to fetch parking lots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
              example:
                success: false
                message: Failed to fetch parking lots
                error: Cannot connect to database
    post:
      summary: Add a new parking lot
      description: Allows an authenticated user to create a new parking lot and its corresponding spots.
      tags:
        - Parking Lots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - primeLocationName
                - address
                - pincode
                - price
                - maxSpots
              properties:
                primeLocationName:
                  type: string
                  example: "Downtown Center"
                address:
                  type: string
                  example: "456 Main Street"
                pincode:
                  type: string
                  example: "560001"
                price:
                  type: number
                  format: float
                  example: 50.00
                maxSpots:
                  type: integer
                  example: 100
      responses:
        "201":
          description: Parking lot added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Parking lot added successfully
        "400":
          description:  User is not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Only admins can add parking lots
        "500":
          description: Server error while adding parking lot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to add parking lot
                  error:
                    type: string
                    example: UNIQUE constraint failed
  /api/parking-lots/{id}:
    put:
      summary: Edit a parking lot
      description: Allows an authenticated user to update the details and size of a specific parking lot.
      tags:
        - Parking Lots
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the parking lot to update
          schema:
            type: integer
            example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                primeLocationName:
                  type: string
                  example: "City Center Garage"
                address:
                  type: string
                  example: "123 Main St"
                pincode:
                  type: string
                  example: "700001"
                price:
                  type: number
                  format: float
                  example: 40.00
                maxSpots:
                  type: integer
                  example: 50
      responses:
        "200":
          description: Parking lot edited successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Parking lot edited successfully
        "400":
          description: Cannot remove occupied parking spots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Cannot remove occupied parking spots
        "404":
          description: Parking lot ID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Parking lot ID not found
        "500":
          description: Server error while editing parking lot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to edit parking lot
                  error:
                    type: string
                    example: Some error details
    delete:
      summary: Delete a parking lot
      description: Deletes a parking lot if none of its spots are currently occupied.
      tags:
        - Parking Lots
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the parking lot to delete
          schema:
            type: integer
            example: 12
      responses:
        "200":
          description: Parking lot successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Parking lot deleted successfully
        "400":
          description: Cannot delete lot due to occupied spots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Cannot delete lot as some spots are currently occupied
        "404":
          description: Parking lot not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Parking lot ID not found
        "500":
          description: Server error while deleting parking lot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to delete parking lot
                  error:
                    type: string
                    example: SQLAlchemy error message
  /api/parking-lots/{lot_id}/spots:
    get:
      summary: Get spots for a specific parking lot
      description: Returns a list of parking spots for the given parking lot ID.
      tags:
        - Parking Lots
      parameters:
        - name: lot_id
          in: path
          required: true
          description: ID of the parking lot
          schema:
            type: integer
      responses:
        "200":
          description: List of parking spots fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  spots:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 12
                        lot_id:
                          type: integer
                          example: 5
                        status:
                          type: string
                          example: available
        "404":
          description: Parking lot not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Parking lot not found
        "500":
          description: Internal server error while fetching parking spots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch parking spots
                  error:
                    type: string
                    example: Database connection error
  /api/parking-lots/{lot_id}/available-spot:
    get:
      summary: Get first available parking spot
      description: Returns the ID of the first available parking spot in the given parking lot.
      tags:
        - Parking Lots
      parameters:
        - name: lot_id
          in: path
          required: true
          description: ID of the parking lot to search in
          schema:
            type: integer
            example: 5
      responses:
        '200':
          description: Successfully found an available parking spot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  spot_id:
                    type: integer
                    example: 12
        '404':
          description: No available parking spot found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: No available parking spot
        '500':
          description: Server error when fetching parking spot
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch parking spot
                  error:
                    type: string
                    example: Database connection error
  /api/parking-lots/search:
    get:
      summary: Search parking lots by location or pincode
      description: Returns a list of parking lots that match the provided location or pincode.
      tags:
        - Parking Lots
      parameters:
        - name: location
          in: query
          required: false
          description: Search term for the prime location name (partial match supported)
          schema:
            type: string
            example: MG Road
        - name: pincode
          in: query
          required: false
          description: Search term for the pincode (partial match supported)
          schema:
            type: string
            example: "560001"
      responses:
        "200":
          description: Matching parking lots fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  lots:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        prime_location_name:
                          type: string
                          example: MG Road
                        address:
                          type: string
                          example: 123 MG Road, Bangalore
                        pincode:
                          type: string
                          example: "560001"
                        price:
                          type: number
                          format: float
                          example: 50.00
                        number_of_spots:
                          type: integer
                          example: 25
        "500":
          description: Internal server error while fetching parking lots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch parking lots
                  error:
                    type: string
                    example: Database connection error
  /api/parking-lots/admin/search:
    get:
      summary: Admin search parking lots by name or pincode
      description: Allows administrators to search for parking lots using either the name or pincode.
      tags:
        - Parking Lots
      parameters:
        - name: parkingLotPincode
          in: query
          required: false
          description: Pincode of the parking lot to search for (partial match supported)
          schema:
            type: string
            example: "560001"
        - name: parkingLotName
          in: query
          required: false
          description: Name of the parking lot to search for (partial match supported)
          schema:
            type: string
            example: Koramangala
      responses:
        "200":
          description: Matching parking lots fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  lots:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        prime_location_name:
                          type: string
                          example: Koramangala
                        address:
                          type: string
                          example: 45, 5th Cross Rd, Bangalore
                        pincode:
                          type: string
                          example: "560034"
                        price:
                          type: number
                          format: float
                          example: 40.00
                        number_of_spots:
                          type: integer
                          example: 30
        "404":
          description: Neither pincode nor name was provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Parking lot ID or name not found
        "400":
          description: User is not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Only admins can search this lot
        "500":
          description: Internal server error while fetching parking lots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch parking lots
                  error:
                    type: string
                    example: Database connection error
  /api/parking-lots/summary:
    get:
      summary: Get user-specific parking lot summary
      description: |
        Returns a summary of all completed reservations for the currently logged-in user, grouped by parking lot location.  
        Shows total visits and total amount spent per location.
      tags:
        - Parking Lots
      responses:
        "200":
          description: Summary fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  lots:
                    type: array
                    items:
                      type: object
                      properties:
                        location:
                          type: string
                          example: Koramangala
                        total_visits:
                          type: integer
                          example: 5
                        total_spent:
                          type: number
                          format: float
                          example: 250.50
        "404":
          description: No data found for user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: No parking lots found
        "500":
          description: Internal server error during summary fetch
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch parking lots data
                  error:
                    type: string
                    example: Database error details
...